+++ backend/schedulerd/roundrobin_cron.go
@@ -80,7 +80,13 @@ func (s *RoundRobinCronScheduler) handleEvent(executor *CheckExecutor, event rin
 			"removing entity from round-robin")
 
 	case ringv2.EventTrigger:
-		s.logger.Info("scheduling check")
+		if len(event.Values) == 0 {
+			s.logger.Error("round robin check scheduled, but no available entities")
+			return
+		}
+		// The ring has produced a trigger for the entity, and a check should
+		// be executed.
+		s.logger.WithFields(logrus.Fields{"agents": event.Values}).Info("executing round robin check on agents")
 		s.mu.Lock()
 		s.schedule(executor, s.proxyEntities, event.Values)
 		s.mu.Unlock()